% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monteriskr-package.R
\docType{package}
\name{monteriskr-package}
\alias{monteriskr}
\alias{monteriskr-package}
\title{monteriskr: Bayesian Demand Modeling and Monte Carlo Risk Analysis for Events}
\description{
Fits Bayesian negative binomial demand models with capacity censoring using brms, simulates posterior predictive demand, and computes Monte Carlo risk metrics for event planning scenarios.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/aryaanswain-13/monteriskr}
  \item Report bugs at \url{https://github.com/aryaanswain-13/monteriskr/issues}
}

}
\author{
\strong{Maintainer}: Aryaan Swain \email{aryaanswain@gmail.com}

Authors:
\itemize{
  \item Aryaan Swain \email{aryaanswain@gmail.com}
}

}
\keyword{internal}
